[flutter_tools] split cache into API and implementation (#79580)
