+++ packages/flutter_tools/test/general.shard/project_test.dart
@@ -814,6 +814,7 @@ void _testInMemory(String description, Future<void> Function() testMethod) {
   transfer(Cache(
     fileSystem: globals.fs,
     logger: logger,
+    artifacts: <ArtifactSet>[],
     osUtils: OperatingSystemUtils(
       fileSystem: globals.fs,
       logger: logger,
@@ -858,6 +859,7 @@ void _testInMemory(String description, Future<void> Function() testMethod) {
         fileSystem: testFileSystem,
         osUtils: globals.os,
         platform: globals.platform,
+        artifacts: <ArtifactSet>[],
       ),
       FlutterProjectFactory: () => FlutterProjectFactory(
         fileSystem: testFileSystem,
